Implemented Monthly & Weekly Report Methods
Initially I had made a method for weekly report and a method for monthly report, however I realized that they were actually quite similar so instead I created a new method that I had take a variable that determines whether the report is weekly or monthly. The report method takes in the current day, month, year and a boolean for weekly or monthly report, then returns a list that contains a count for each emoji used, which represents the summary of your week or month, as the emojis are a more general representation of how you were feeling so compiling them together should be representative.

Also added test cases for a week with no entries and a week with one entry