Weekly report
added weekly report. The report method takes in the current day, month, year then returns a list that contains a count for each emoji used, which represents the summary of your week, as the emojis are a more general representation of how you were feeling so compiling them together should be representative.

Also added test cases for a week with no entries and a week with one entry