Monthly Report
added monthly report. The report method takes in the current day, month, year then returns a list that contains a count for each emoji used, which represents the summary of your month, as the emojis are a more general representation of how you were feeling so compiling them together should be representative.

Also added test cases for a month with no entries and a month with one entry